mkosi: add support for running arbitrary build logic within the image setup
If an executable script "mkosi.build" exists in the local directory, the
build logic will now build the image twice. First, it will build an
image that the working directory and "mkosi.build" is copied into (under
the assumption the tool is called from a source tree of a project). It
then runs that script inside the image via nspawn, and collects any
output it places in a directory passed as $DESTDIR. Then, it throws the
first image away, builds a new one, this time avoiding any source tree or
script copying. However, as final step of the image preparation the tree
from $DESTDIR is copied into the image.

The first image may optionally contain additional distro packages, on
top of the packages the final image will contain. This is useful for
adding in development headers and suchlike.

The suggested usecase for this is the following: let's say you develop a
project, and want to quickly build an OS image from it, with your
current development tree as source tree for the build, and you want to
build to take place inside an image that resembles the ultimate image to
generate as closely as possible. Hence we first build an image with all
development headers for compiling your project. Then artifacts it
generates are then included in the more minimal second image, which is
the result.