In a git repo, do not copy files that git itself would ignore
The file list to copy is generated using git ls-files --other --cached,
i.e. anything in git ignore lists is ignored.

Without that, mkosi chokes on the test/sys directory in systemd source
tree that is generated as part of tests. I think the new behaviour is
a good default, since most of the times people are going to be running
this from a git tree anyway.

Also skip .git directory by default.

This requires mkosi.build to be updated to remove 'git clean …'.

Fixes #22.