Allow setting root password in the image
Those images are supposed to be used locally for testing, so it
should be as easy to log in as possible.

As a special case, --password= completely deletes the password.
In other words, it's not possible to set an empty password.

Note: the obvious thing of calling /bin/passwd does not work because
audit or some other thing trips it up.

Thanks to Lucas Werkmeister for the crypt.crypt() line!