mkosi: set GPT read-only flags properly
This makes sure the --read-only setting has an effect on the GPT
read-only flags of the root partition.

Note that we generate the sfdisk lines in a format that works around an
sfdisk bug with older version: we place the partition name last in each
line, as we need to enclose it in "", but the sfdisk parser is confused
by it otherwise. (See: https://github.com/karelzak/util-linux/issues/380)