mkosi: rename SHA256SUM → SHA256SUMS
This was an unfortunate typo I unfortunately didn't notice earlier. The
file was really supposed to be created as SHA256SUMS rather than
SHA256SUM, in order to follow Ubuntu's naming of the checksums of OS
images.  Moreover systemd's importd tool will actually look for
SHA256SUMS rather than SHA256SUM.

Hence, let's rename this now, even if this is to some level a compat
break. But as nobody noticed this earlier I figure people didn't care so
far.