fix breakage when making directory format in current directory
Making an osi of type "directory" when using the default path (the
current working directory) causes the build step to fail. This occurs
because mkosi attempts to copy the .mkosi-* directory into itself and
causes an infinite recursion.

This also resolve an issue where directories that don't contain any
tracked files are copied to the build directory (e.g. the .git
directory).