mkosi: support reusing builddirs between subsequent mkosi runs
With this change, mkosi will look for "mkosi.builddir" when invoked.
When this directory exists it is mounted into the build container and
may be used as location for out-of-tree builds, which can be shared
between multiple invocations to speed up the building process.

This is particularly useful when used in conjunction with incremental
building (-i), as with this in place building OS images using mkosi is
only a little slower than just building the project natively on the host
now.

"mkosi clean -f" will empty the builddir if it exists now too, in
addition to the incremental image trees.