mkosi: add two new verbs "boot" and "shell"
Both verbs are very similar: they build the image similar to "mkosi
build" (except if it exists already) and then invoke "systemd-nspawn"
interactively on the output image. "boot" will pass "--boot" to nspawn,
while "shell" doesn't.

The command may be combined with "-f" as usual, to force a rebuild
before invoking nspawn on the result.

With this in place getting to a test shell from a current git checkout
of a project like systemd is as easy as typing "mkosi shell".

In a later PR we should really add a "mkosi vm" command or so, which
permits booting up the image in qemu properly finding UEFI and kvm
support for it, and so on.