Fix btrfs_subvol_delete by deleting recursively
`btrfs subvol delete` errors out when the subvolume still contains
anything so we have to delete the contents first. However the subvolumes
we have to delete usually themselves contain subvolumes so we can't just
use `rm -r`. What we do instead is go through the directory depth first
using `find` and then try to delete each entry with `rm -d` first and if
that fails check if the entry is a directory and if so assume it's a
subvolume and fall back to `btrfs subvol delete`.