mkosi: Recursively delete btrfs subvolumes
`btrfs subvol delete` can only delete a subvolume if it is not read-only
and if it does not contain another subvolume.

`btrfs subvol list -o` prints the subvolumes below a certain path, but
only gives the paths relative to the filesystem mount point for those
subvolumes. To compute the paths of the children, we need the relative
path of the parent subvolume given by `btrfs subvol show`.

Read-only subvolumes also have to be writeable again before being
deleted. We unconditionally unset the readonly property on a subvolume
before deletion.