mkosi: add "qemu" verb
This adds a new "mkosi qemu" verb that is like "mkosi boot" but invokes
things in a QEMU environment.

This tries to find the native qemu implementation, with a couple of
fallbacks if none is found, always opting for KVM acceleration.

This also tries to automatically discover the UEFI firmware blobs. For
now only the location where the files are on Fedora is encoded in the
search path, but this can be updated for other distributions easily
later on.

Fixes: #179