mkosi: create builddir if configured but missing
If a builddir is explicitly configured but missing, automatically create
it.

This follows the logic already implemented for mkosi.output/ and
mkosi.cache/.

This is particularly useful when storing mkosi settings in a git repo,
as git normally doesn't allow us to store empty directories, and hence
we cannot make use of auto-discovery of mkosi.output/ and mkosi.cache/.
By configuring these paths explicitly in mkosi.default however, we can
work around this, as in that case we'll create the directories if
needed.