mkosi: btrfs_subvol_delete with mounted subvolumes
Recursively deleting a subvolume does not work if an ancestor subvolume
is explicitly mounted (e.g. with `mount -o subvolid=<ID>`).

In such a case, the paths returned by `btrfs subvol list -o` are
specific to the BTRFS filesystem and have no relation with the mount
point.  An incorrect path is then constructed for the children and their
deletion will fail.

Instead, we retrieve the ID of the subvolumes below the current one and
query their path relative to the current subvolume with `btrfs subvol
show -r`.