mkosi: try hard to cleanup the workspace
The workspace is a `tempfile.TemporaryDirectory` and its cleanup fails
if it cannot delete a file or directory contained in the workspace.

This is the case when an exception occurs during the build of a BTRFS
subvolume image. `tempfile.TemporaryDirectory` cannot delete the
subvolumes and the workspace remains.

The workspace is now an instance of `Workspace`, a class that wraps a
`tempfile.TemporaryDirectory`. When the workspace instance is deleted,
its contents are first deleted with unlink_try_hard and the temporary
directory can finally be cleaned up.

Fixes #182