Fix copy of symlinks
When copying files we should not be following the symlinks. This became
a problem in kmod after removing the git-clone method to copy files
because we have intentionally broken symlinks there:

  File "/usr/bin/mkosi", line 170, in open_close
      fd = os.open(path, flags | os.O_CLOEXEC, mode)
      FileNotFoundError: [Errno 2] No such file or directory:
      '/home/lucas/p/kmod/testsuite/rootfs-pristine/test-loaded/sys/module/btusb/drivers/usb:btusb'
      make: *** [Makefile:3009: mkosi] Error 1

Although it's worse when the symlink is broken, it would silently succeed when
it's not but producing an invalid output.