Support copying git submodule files
Besides getting the files from the git directory, iterate through each
submodules to copy them over to the build image as well. We had some
(better) alternative, but each of them failing to provide what we need:

    - Recent versions for git-ls-files learned a --recurse-submodules
      but they are not support together with --others

    - git-submodule foreach --recursive would allow us to git-ls-files
      inside each submodule, however there's no easy way to control from
      which submodule the command is printing the files in order to
      prepend the submodule path

So for now we live with getting the list of submodules from
`git submodule status --recursive` and calling git-ls-files on each of
them.