fedora: cut down on the installation set size
By installing glibc-minimal-langpack we avoid the installation of glibc-all-langpacks
(glibc requires glibc-langpack (provided by all langpacks) and suggests
glibc-all-langpacks (so that is the default when there is nothing else to
satisfy the require). By forcing a different langpack package, we avoid the
installation of any others.

git-core is similar to git, but does not pull in some lisp and perl stuff.

kernel-core is similar to kernel, but does not include some fancy drivers.