Allow to run qemu >= 2.12
While trying to call "mkosi qemu" on a Clear Linux host it reported

    qemu: could not load PC BIOS '/usr/share/qemu/OVMF_CODE.fd'

Cross checking with my Fedora host the only real difference I see is the
qemu version. Changing it to pass a pflash drive directly makes it work
and also make it explicit that the it is a readonly device.