Rename --format members from raw_* to gpt_*
They are not "raw" in any meaningful way, since the filesystems are
packaged in a disk image with a gpt partition. I know that "raw" comes
from qemu, where it used to mean e.g. not-qcow2, but since we don't
produce qcow2 or any of the other formats, here it is just misleading.

Also, use auto() to number OutputFormat members and don't show compat names
in help output.

v2:
- add .__repr__() and .__str__() and .from_string() to make the --help
  output message and the error message when an unknown type is given nicer.