Remove yum/dnf/rpm metadata after installation
Dnf data is removed if dnf is not installed into the image, and similarly
for the other tools.

I didn't implement this functionality for Debian/Ubuntu/Arch, because I
don't know enough about those distros, but is should be easy to add.

The arguments that are passwd to build_image() and friends are made
keyword-only. There's now three arguments, and this way there's less
chance of mistake.

v2:
- When deciding whether to keep dnf or rpm data, check if path exists

Checking if the dnf/rpm packages are in the requested list is nice,
but inadequate. For example, somebody could specifyg a @group or just
another package that pulls those in. Checking if the executables are
present in the image would work in those cases too.