Fix possible None value in prepare_swap
Mypy reports that partition() may return None and break the
requirements of subprocess.run():

    mkosi:525: error: List item 2 has incompatible type "Optional[str]";
    expected "Union[bytes, str, _PathLike[Any]]"

There are plenty of issues like so please treat this patch as an RFC. I
think the best way forward is to reduce the number of optional values to
the point where such checks are not needed. The generic nature of many
helpers (looking at the args and returning None if some condition
doesn't hold) amplifies this problem.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>