Fix broken bootable images
We need to bring in systemd-udev to the packages list if the image is
bootable, not if we are going to run the build script.

$ sudo mkosi -d fedora -r 29 -t raw_ext4 -b
...
FileNotFoundError: [Errno 2] No such file or directory: '/var/tmp/mkosi-a9r063jq/root/usr/lib/systemd/boot/efi/systemd-bootx64.efi'

This could maybe be in make_rpm_list(), but the packages names for
Mageia are different than the ones for Centos/Fedora.