Add support to boot on bios
This implements boot on systems with legacy bios rather
than UEFI. It works by installing grub-pc and systemd-boot
side by side. The former is responsible for bios boot while
the latter continues to be the bootloader for UEFI.

The resulting image, that is currently working only on Fedora,
is capable of booting on both systems. Generate the image with:

    $ sudo mkosi -d fedora -r 29 -t raw_ext4 -b --boot-protocols uefi bios

Boot with UEFI:

    $ sudo qemu-system-x86_64 -machine accel=kvm -smp 2 -m 1024 \
        -drive if=pflash,format=raw,readonly,file=/usr/share/edk2/ovmf/OVMF_CODE.fd \
        -drive format=raw,file=image.raw

Boot with bios:

    $ sudo qemu-system-x86_64 -machine accel=kvm -smp 2 -m 1024 \
        -drive format=raw,file=image.raw

Lucas: This is a cleanup of the implementation done by Rafael
and allowing to have the hybrid approach rather than mandating one
or the other.