Chown output files to user calling sudo
In most of the cases we are using sudo to elevate privileges and be able
to do all operations we need. The final files created by mkosi though
should better be owned by the original user so the user may decide what
to do with it like copying it somewhere else, starting qemu, etc.