mkosi: pass $WITH_NETWORK= to mkosi.build script
Let's tell the build script whether it is running with networking or
not. I don#t have an immediate usecase for this, but it's more
systematic given how --with-docs and --without-tests is currently
handled and result in environment variables, bus --with-network is not.

A usecase might be to generate the `-nonet` option of `xsltproc` from
this option, but that's purely theoretical.