man: add a man page
This adds a pretty comprehensive man page as a MarkDown document.

This does not contain any build system hookup, simply because I couldn't
figure out how to teach Python's setuptools to invoke external programs
as part of the build process.

Currently a manual step is necessary, to convert the man page into an
actual man page:

        # pandoc mkosi.md -s -t man > mkosi.1

The man page is based on the old README.md but substantially extended.

Fixes: #210