Fix setting of boolean parameters in mkosi.default
Some boolean parameters are currently not taken into account when specified
in `mkosi.default`.
The reason is that, when parsing `mkosi.default`, in `process_setting`,
`args` parameters are set only if their previous value is `None` so that
parameters passed on the command line have the priority over the content
of `mkosi.default`.
But, when a parameter defined in `parse_args`, has `action='store_true'`,
its default value is `False`.

So, let’s put `None` default value in `parse_args` so that parameters
not specified on the command line can be overridden in `process_setting`.