Optionally include the .git directory with the files it contains
When copying files in from git, using copy-git-cached ignores the files
in the .git/ directory. This introduces a copy-git-more option that also
pulls those files in. This is useful for builds that use `git describe`
or another git command for their version number or tests. With this
feature they can now query it.