mkosi: incremental builds are not supported for squashfs or minimized file systems
This was never supported for squashfs, but failed badly. Let's generate
a nice error instead.

(The thing with these builds are that while we build these images we
don't mount the final file systems inside the image file. Instead we
mount a temporary directory in /var/tmp that is packed up only much
later. This makes caching harder, since it's not just one dir or file we
need to cache but the image *and* this directory. We can add support for
this eventually, but for now let's just friendly decline operation. It's
not too limiting, as in particular minimization is mostly an
optimization that is only useful on final builds, not on test builds so
much, while incremental mode is useful for test builds but not so much
for final builds.)