add first boot parameter
When generating cloud images or other images that should
usesystemd-firstboot, we want to make sure that /etc/machine-id is
not existent. Creating an empty /etc/machine-id file is not enough,
because systemds `ConditionFirstBoot` will only check for the file, not
its content. Furthermore systemd will generate a new /etc/machine-id file
if the /etc/machine-id file is empty.

Signed-off-by: Christian Rebischke <chris@nullday.de>