Gracefully handle BIOS + --minimize/gpt_squashfs combo
Using grub on generated_root() images currently has problems as there is
a circular dependency between the root partition wanting to have grub
installed before being generated and grub wanting the root partition to
be generated before being installed.

It might be possible to get this working but until we do it's better to
explicitly fail when users try to do this.