allow optional override of GPT first-lba header
Usage of this feature is totally optional and the defaults are actually
preferred, which should set the First (usable) Logical Byte Address to
sector 2048, leaving a 1MiB gap before the first partition.
However there might be a reason to override this, e.g. if the created
image is being altered later on by tools which need a different
first-lba value. One example would be the `prl_disk_tool` utility from
the Parallels virtualization suite.